refactor(Dashboard): 탭 기능별 컴포넌트 모듈화
- **목적**: DashboardView의 코드 복잡도를 낮추고, 향후 백엔드 연동 및 기능 확장을 용이하게 하기 위해 리팩토링을 진행했습니다.
- **변경 사항**:
    - \DashboardView\는 각 탭을 불러오는 컨테이너 역할만 하도록 수정
    - \DashboardTab\, \LinesTab\, \PredictionTab\, \InventoryTab\ 컴포넌트 신규 생성
    - 각 컴포넌트는 독립적인 UI와 로직을 갖도록 분리
- **기대 효과**:
    - 컴포넌트 단위의 명확한 역할 분리
    - 코드 추적 및 디버깅 용이성 향상
    - 백엔드 API 연동 시 작업 범위 명확화
